requirements.txt
flask==3.0.3
flask-cors==4.0.1
groq==0.9.0
requests==2.32.3
feedparser==6.0.11
pdfplumber==0.11.4
python-docx==1.1.2
